iloveyouuu

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User"
 
 export const home = async(req, res) => {
     const videos = await Video.find({}).sort({createdAt:"desc"});
@@ -7,7 +8,7 @@ export const home = async(req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.render("404", {pageTitle:"Video not found."});
     }
@@ -16,9 +17,13 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", {pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Edit ${video.title}`, video});
 };
@@ -43,13 +48,20 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async(req, res) => {
+    const { user: {_id}, } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner:_id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.dinfBuId(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.direct("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -61,14 +73,22 @@ export const postUpload = async(req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findBiId(id);
+    if(!video){
+        return res.render("404", {pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
 
-export const search = async(req, res) => {
+export const search = async (req, res) => {
     const { keyword } = req.query;
     let videos = [];
-    if (keyword) {
+    if(keyword){
         videos = await Video.find({
             title: {
                 $regex: new RegExp(keyword, "i"),
